Added sort implementation based on https://github.com/ccharly/brigand/commit/97f76c16ef6979cb3b990efd5fa3241f112c7baf . Rewrote partition implementation based on my filter implementation's structure.
Note that @ccharly 's partition implementation returned its two sets reversed from the order than the standard library and other languages use; modified to be consistent with the others.

Only tested on MSVC 2015 RTM; blame the bizarre code structure on the myriad hoops that had to be jumped through to avoid ICEs.